refactor code, now the pid controller is its own package and can accept joint position commands
