+++ README.md
@@ -8,6 +8,7 @@ This repository contains a ROS2 package for simulating and visualizing a UR3 rob
 - [Usage](#usage)
   - [Launching the Simulation](#launching-the-simulation)
   - [Visualizing with RViz2](#visualizing-with-rviz2)
+  - [PID control](#pid-controller)
 - [Contributing](#contributing)
 - [License](#license)
 
@@ -67,6 +68,25 @@ Rviz2 is used to render the robot inside the simulation. To launch rendering, us
 ```bash
 ros2 launch src/ros2_sim_ur3_description/launch/simulation.launch.py
 ```
+
+## PID controller
+
+The PID controller package is responsible for turning desired joint positions into torque values for the joints. Currently, the gains are not perfectly tuned yet, but the controller works properly.
+
+TODO: add back optional tuning code for the PID controller
+
+```bash
+ros2 launch src/ros2_sim_pid_controller/launch/pid_controller.launch.py
+```
+
+The PID controller can be tested via the terminal using:
+
+```bash
+ros2 topic pub /desired_positions sensor_msgs/msg/JointState "{header: {stamp: {sec: 0, nanosec: 0}}, name: [shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'], position: [0.2, 0.2, 0.0, 0.0, 0.0, 0.0]}"
+```
+
+This command sends the joints to [0.2, 0.2, 0.0, 0.0, 0.0, 0.0].
+
 
 # Contributing
